Fixed errors of the home page, added functionality
Added a transition to the telegram channel of IIKS, INTEL, the MEPhI mini football club, to the official MEPHI website from the main page, corrected the position of the image on the screen. Then it would be nice to make a transition to the sources not immediately, but to go to another page where some information about the source will be presented, as well as a button that will allow you to go to the source.